prevent overwriting old backup files

+++ setup.bash
@@ -14,16 +14,24 @@ dirList=( .vim )
 # backup old dirs and files
 for f in "${fileList[@]}"; do
 	if [[ -f "$f" && ! -L "$f" ]]; then
-		echo "Renaming $f to $f.old ..."
-		mv $f $f.old
+		if [[ -f "$f.old" ]]; then
+			echo "$f.old file already exists! Remove it and try again."
+		else
+			echo "Renaming $f to $f.old ..."
+			mv $f $f.old
+		fi
 	fi
 	echo "Creating symlink $f pointing $cwd/$f ..."
 	ln -s -f $cwd/$f $f
 done
 for f in "${dirList[@]}"; do
 	if [[ -d "$f" && ! -L "$f" ]]; then
-		echo "Renaming $f to $f.old ..."
-		mv $f $f.old
+		if [[ -d "$f.old" ]]; then
+			echo "$f.old directory already exists! Remove it and try again."
+		else
+			echo "Renaming $f to $f.old ..."
+			mv $f $f.old
+		fi
 	fi
 	echo "Creating symlink $f pointing $cwd/$f ..."
 	ln -s -f $cwd/$f $f
